// src/types/audit.ts
// Audit event types - keep in sync with Prisma schema
export enum AuditEventType {
  LOT_CREATED = 'LOT_CREATED',
  LOT_UPDATED = 'LOT_UPDATED',  
  LOT_DELETED = 'LOT_DELETED',
  AUCTION_STARTED = 'AUCTION_STARTED',
  AUCTION_CLOSED = 'AUCTION_CLOSED',
  AUCTION_CANCELLED = 'AUCTION_CANCELLED',
  BID_PLACED = 'BID_PLACED',
  BID_WITHDRAWN = 'BID_WITHDRAWN',
  TRANSACTION_CREATED = 'TRANSACTION_CREATED',
  TRANSACTION_UPDATED = 'TRANSACTION_UPDATED',
  SETTLEMENT_CREATED = 'SETTLEMENT_CREATED',
  SETTLEMENT_COMPLETED = 'SETTLEMENT_COMPLETED',
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_DELETED = 'USER_DELETED',
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILED = 'LOGIN_FAILED',
  PERMISSION_DENIED = 'PERMISSION_DENIED',
}